move from demo to catalog package
