+++ src/main/java/com/wine/to/up/catalog/service/domain/enums/Color.java
@@ -1,0 +1,7 @@
+package com.wine.to.up.catalog.service.domain.enums;
+
+public enum Color {
+    WHITE,
+    PINK,
+    RED
+}
