+++ src/main/java/com/wine/to/up/catalog/service/domain/enums/Sugar.java
@@ -1,0 +1,8 @@
+package com.wine.to.up.catalog.service.domain.enums;
+
+public enum Sugar {
+    DRY,
+    SEMI_DRY,
+    SEMI_SWEET,
+    SWEET,
+};
