+++ src/main/java/com/wine/to/up/catalog/service/messaging/serialization/EventSerializer.java
@@ -1,0 +1,17 @@
+package com.wine.to.up.catalog.service.messaging.serialization;
+
+import com.wine.to.up.demo.service.api.message.KafkaMessageSentEventOuterClass.KafkaMessageSentEvent;
+import org.apache.kafka.common.serialization.Serializer;
+
+/**
+ * Serializer for {@link KafkaMessageSentEvent}
+ */
+public class EventSerializer implements Serializer<KafkaMessageSentEvent> {
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public byte[] serialize(String topic, KafkaMessageSentEvent data) {
+        return data.toByteArray();
+    }
+}
