+++ src/main/java/com/wine/to/up/catalog/service/service/DiscountInfoService.java
@@ -1,0 +1,21 @@
+package com.wine.to.up.catalog.service.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class DiscountInfoService {
+    public Object createDiscountInfo(Object object) {
+//        System.out.println("Hello, Discount!");
+        return null;
+    }
+
+    public Object updateDiscountInfoById(String id, Object obj) {
+
+        return null;
+    }
+
+    public Object getDiscountInfoById(String id) {
+
+        return null;
+    }
+}
