+++ src/main/java/com/wine/to/up/catalog/service/service/GrapeManagerService.java
@@ -1,0 +1,36 @@
+package com.wine.to.up.catalog.service.service;
+
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+//@RequiredArgsConstructor
+public class GrapeManagerService {
+
+//    @Autowired
+//    public WineManagerService wineManagerService;
+
+    public Object getGrapeById(String id) {
+
+        return null;
+    }
+
+    public List<Object> getAllGrapes() {
+
+        return null;
+    }
+
+    public Object updateGrape(String id, Object obj) {
+
+        return null;
+    }
+
+    public Object createGrape(Object obj) {
+
+        return obj;
+    }
+}
