--- src/main/java/com/wine/to/up/demo/service/domain/entities/Brand.java
@@ -1,18 +1,0 @@
-package com.wine.to.up.demo.service.domain.entities;
-
-import lombok.Data;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "Brands")
-@Data
-public class Brand {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "brandID", nullable = false)
-    private int brandID;
-
-    @Column(name = "brandName")
-    private String brandName;
-}
