--- src/main/java/com/wine/to/up/demo/service/domain/enums/Color.java
@@ -1,7 +1,0 @@
-package com.wine.to.up.demo.service.domain.enums;
-
-public enum Color {
-    WHITE,
-    PINK,
-    RED
-}
