--- src/main/java/com/wine/to/up/demo/service/domain/enums/Sugar.java
@@ -1,8 +1,0 @@
-package com.wine.to.up.demo.service.domain.enums;
-
-public enum Sugar {
-    DRY,
-    SEMI_DRY,
-    SEMI_SWEET,
-    SWEET,
-};
